requirements.txt
pandas~=2.2.2
folium~=0.17.0
matplotlib~=3.9.0
scipy~=1.13.1
python-dotenv
pybind11>=2.12
opencv-python~=4.10.0.84
numpy<2
numexpr==2.8.4
bottleneck==1.3.6
ttkbootstrap
